ARCompact: implement min and max instructions
Following a comment on GitHub
(https://github.com/NationalSecurityAgency/ghidra/pull/3006#issuecomment-833600575),
use an expression similar as the one in `PIC/data/languages/PIC24.sinc`
to implement instructions `min` and `max`.

While at it, fix the computations of the condition flags in `min.f` and
`max.f`, which were wrong.